RDKB-6273: User redirected to captive portal even after customization
Reason for change:
Through SUSI, when user activates account and enters WiFi personalization parameter,sometime user is not coming out of captive portal.
This was happening due to a synchronization issue when notification of parameters are raised in RDKB.

Fix is to raise notification using a single thread and verify if values are really changed.
Test Procedure:
1) loaded image with fix in XB3
2) Activated unit through SUSI
3) Verifed user is out of captive portal after the activation flow
4) Verified captive portal functionality through local captive portal

Change-Id: Ifce3286c350b63e09f7080103a1e13a9c2d14cf1
Risks: None
Signed-off-by: achelakkode<anoopnair_c@infosys.com>