RDKB-7775: Remove banner from dropbear daemon
Reason for change:
In RDKB devices, as dropbear is initialized with a banner whenever there is a rsync from ATOM to ARM during dca execution, banner is printed to ArmConsoleLog.txt.
 This is redundant and in certain field devices this is causing log upload frequently.

Test Procedure:
Verified in ArmConsolelog and AtomConsole log files, no banner content is printed when SCP/rsync/dca is in progress.

Change-Id: Ic6766c97767c9828e49191b66ab49db6ef37a5ea
Risks: None
Signed-off-by: achelakkode<anoopnair_c@infosys.com>