RDKB-29168 RDKB-29102: RBus returns dummy as CompName whereas DBus returns Error
Reason for change: when the NameSpace discovery was initiated, the DBus returns Error saying NameSpace Not Supported but the RBus was returning DUMMY as Component Name. This leads the SNMP or PAM Component to make a Query and its ends up with Error message.
This is not harming anything; as Dbus doesnt even make GET as the query has failed; whereas Rbus makes the query and fails. But to avoid unnecessary socket call and to avoid error message in the log file, we compromise this.
Test Procedure: As per RDKB-29168
Risks: Medium
Signed-off-by: Karunakaran A <karunakaran_amirthalingam@cable.comcast.com>

RDKB-29102 : Return Component NameSpace for SNMP

Reason for change: The SNMP component currently has hardcoded way to set/get the params from the same component only as it is designed on top of DBUS/CCSP.
RBus doesnt need this. But since we dont want to restructure SNMP, this will give better handle.
Test Procedure: As per RDKB-29182
Risks: Low
Signed-off-by: Karunakaran A <Karunakaran_Amirthalingam@cable.comcast.com>

Change-Id: I7b712c643c1017433fb692b199d527c3b974cad5