RDKCOM-3049: RDKBDEV-971: validate Notification when bTargetIsObject is FALSE
Reason for change:
DslhWmpDatabase: validate Notification when bTargetIsObject is FALSE
Validate the Notification value for the parameter in the case that
bTargetIsObject is FALSE in DslhWmpdoMpaSetParameterAttributes().
A similar check already exists for bTargetIsObject is TRUE (keep that
code unchanged but reformat slightly for clarity).
A slightly different Notification value check already existed in the
code for bTargetIsObject is FALSE but has been commented out - keep
that code as a reference. It may be that the final solution should be
a combination of the two checks?
Test Procedure:
 - SetParameterAttributes for "Device.WiFi.SSID.10001.SSID" works.
 - SetParameterAttributes for "Device.WiFi.SSID.10001.Stats.BytesReceived" is rejected with:
   {'faultcode': 'Server', 'code': '9009', 'detail': 'Notification request rejected'}.
Risks: None.
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

(cherry picked from commit d6fb38c8cefefa7c79909bccadfbfb97c74982dd)
Change-Id: Ib22013c20d23e9f04cf871eed12fa5439f03f965