RDKCOM-2812: RDKBDEV-856: only zero pTempBuf if it's going to be used
Reason for change:
only zero pTempBuf in DslhObjcoGetParamValueByName() if it's going to be used
The pTempBuf buffer is 1k bytes, so zeroing it on every call to
DslhObjcoGetParamValueByName() is wasteful if the buffer is only used
for objects which return a string or byte array.
Test Procedure: Sanity.
Risks: None.
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

(cherry picked from commit fc950cfe981ea3ee3dbbfc5dfa4f08616c8d05bf)
Change-Id: I4ad2c716d0d6094a29dc21c179f83d23f2b7a6c3