RDKBDEV-856: only zero pTempBuf if it's going to be used
Reason for change:
only zero pTempBuf in DslhObjcoGetParamValueByName() if it's going to be used
The pTempBuf buffer is 1k bytes, so zeroing it on every call to
DslhObjcoGetParamValueByName() is wasteful if the buffer is only used
for objects which return a string or byte array.

Test Procedure: Sanity.
Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: Ib120bbd05ce33ab242d03ad7cd6ec5f037ebcc53